requirements.txt
bson==0.5.8
dataclasses==0.7
Django==2.2.12
django-cors-headers==3.2.1
djongo==1.3.1
dnspython==1.16.0
mysql-connector-python==8.0.19
protobuf==3.6.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
